Created GameLoop.java as per specified design.
